lcms base class
