+++ requirements-dev.txt
@@ -2,7 +2,7 @@ pre-commit
 sphinx
 sphinx_rtd_theme
 pytest
-nmdc-schema~=11.4.0
+nmdc-schema~=11.6.1
 requests~=2.32.3
 requests-oauthlib~=2.0.0
 linkml-runtime~=1.8.2
